mend
!

+++ src/core/base/config/config.go
@@ -1,7 +1,7 @@
 package config
 
 const (
-	MODULE_CODE int16  = 100    // Глобальный код модуля
+	MODULE_CODE int    = 100    // Глобальный код модуля
 	MODULE_NAME string = `base` // Имя модуля
 	PAGE_ITEM   int    = 4
 	PAGE_STEP   int    = 10
